JBPM-7834: Provide OpenShiftStartupStrategy - Enhancements related to RHPAM templates and others
